Fix audio cutoff issues
This fix the audio cutoff issues when building in release mode

Explaination from @garrettjoecox
The result of osAiSetFrequency was being used later on to multiply the length an audio sequence would play. In debug mode this stub was returning 1, whereas in release mode this would return whatever was passed in as the freq argument.

Co-Authored-By: Garrett Cox <garrettjcox@gmail.com>